news_portal/requirements.txt
Django
psycopg2
Pillow
djangorestframework
newsapi-python

